requirements.txt
groq 
opencv-python 
google-genai
ultralytics 
huggingface_hub 
supervision
pillow
numpy
dotenv
huggingface_hub[cli]
transformers
geocoder
phidata
openai
exa_py
fastapi
uvicorn
python-multipart
requests